% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCode.R
\name{extractCode-defunct}
\alias{extractCode-defunct}
\alias{extractCode}
\title{Extract Code From Rpubs Article}
\usage{
extractCode(url, path = NULL, output = FALSE, ...)
}
\arguments{
\item{url}{Character. URL of RPubs article, e.g. url = "http://rpubs.com/aephidayatuloh/sendgmail".}

\item{path}{Character. File name for the extracted code as R script, e.g. \code{code.R}.}

\item{output}{Logical. Should extraction include output of the code? Default to \code{FALSE}, means only R script will be extracted.}

\item{...}{Other parameter for defunct.}
}
\value{
vector
}
\description{
(Deprecated) Extract code from an Rpubs <https://rpubs.com/> article. These functions still work but will be removed (defunct) in the next version.


\itemize{
 \item \code{\link{extractCode}}: This function is deprecated, and will
 be removed in the next version of this package. Use \code{rpubs_code()} instead.
}
}
\details{
If \code{path = NULL} then the extracted script will be print on console or as vector if you assign to an object. One code block is one element of vector.
}
\examples{
\donttest{
extractCode(url = "http://rpubs.com/aephidayatuloh/sendgmail", path = NULL, output = FALSE)
}


}
