% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubs_ly.R
\name{rpubs_ly}
\alias{rpubs_ly}
\title{Extract Code From RPubs Article}
\usage{
rpubs_ly()
}
\description{
Get the information of recently published artiles from an Rpubs homepage.
}
\examples{
rpubs_ly()



}
